requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
pandas
groq
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy
gunicorn

